refactor: 응집도를 위해 훅을 사용하는쪽으로 이동
