+++ shared/components/input/use-input-interaction.ts
